requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
